# config/api.yaml
openapi: 3.0.3
info:
  title: Lethean VPN OpenAPI
  description: |-
    This is Lethean VPN API definition
  license:
      name: GPLv3
  version: "1.0"

paths:
  /api/vdp:
    get:
      summary: Get VDP
      description: Get VPN Description Protocol containing all Spaces and Gateways
      operationId: getVdp
      parameters:
        - in: query
          name: checkOnly
          schema:
            type: boolean
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vdp'
    post:
      summary: Post new VDP to server
      description: Post new VDP to server
      operationId: pushVdp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vdp'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vdp'
        '443':
          description: Error in VDP

  /api/connect:
    post:
      summary: Prepare and connect to Space using Gate
      description: Prepare and connect to Space using Gate
      operationId: Connect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionRequest'
      responses:
        '200':
          description: Connected
        '400':
          description: Bad request
        '402':
          description: Waiting for payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionInfo'
        '460':
          description: Unknown/bad space
        '461':
          description: Unknown/bad gate
        '462':
          description: Cannot use this space/gate combination
        '422':
          description: Bad signature
        '500':
          description: Server error

components:
  schemas:

    Provider:
      type: object
      properties:
        filetype:
          type: string
          enum:
            - LetheanProvider
        version:
          type: string
          enum:
            - "1.0"
        providerid:
          type: string
        name:
          type: string
        description:
          type: string
        ca:
          type: array
          items:
            type: string
        manager-url:
          type: string
        spaces:
          type: array
          items:
            type: string
      required:
        - filetype
        - version
        - providerid
        - name
        - manager-url
        - spaces
        - ca

    Providers:
      type: array
      items:
        $ref: '#/components/schemas/Provider'

    Gate:
      type: object
      properties:
        filetype:
          type: string
          enum:
            - LetheanGateway
        version:
          type: string
          enum:
            - "1.0"
        providerid:
          type: string
        gateid:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: object
          properties:
            per-day:
              type: number
        internal:
          type: boolean
          default: false
        type:
          type: string
          enum:
            - wg
            - http-proxy
            - socks-proxy
            - daemon-rpc-proxy
            - daemon-p2p-proxy
            - wg-tcp
            - ssh
            - tls-ssh
        wg:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            tcp_port:
              type: integer
            public_key:
              type: string
        http-proxy:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            tls:
              type: boolean
        socks-proxy:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            tls:
              type: boolean
        daemon-p2p:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            tls:
              type: boolean
            coin:
              type: string
              default: lethean
              enum:
                - lethean
                - monero
        daemon-rpc:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            tls:
              type: boolean
            coin:
              type: string
              default: lethean
              enum:
                - lethean
                - monero
        ssh:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            rsa-server-public-key:
              type: string
            ecdsa-server-public-key:
              type: string
            username:
              type: string
            client-private-key:
              type: string
            tls:
              type: boolean
          required:
            - host
            - port
            - username
            - client-private-key
        spaces:
          type: array
          items:
            type: string
        gate-types:
          type: array
          items:
            type: string
            enum:
              - wg
              - http-proxy
              - socks-proxy
              - daemon-rpc
              - daemon-p2p
              - wg-tcp
              - ssh
              - ssh-tls
      required:
        - gateid
        - name
        - providerid
        - version
        - filetype
        - type
        - spaces

    Gates:
      type: array
      items:
        $ref: '#/components/schemas/Gate'

    Space:
      type: object
      properties:
        filetype:
          type: string
          enum:
            - LetheanSpace
        version:
          type: string
          enum:
            - "1.0"
        providerid:
          type: string
        spaceid:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: object
          properties:
            per-day:
              type: number
        ips:
          type: array
          items:
            type: string
        dns_servers:
          type: array
          items:
            type: string
      required:
        - spaceid
        - providerid
        - filetype
        - version

    Spaces:
      type: array
      items:
        $ref: '#/components/schemas/Space'

    Vdp:
      type: object
      properties:
        filetype:
          type: string
          enum:
            - VPNDescriptionProtocol
        version:
          type: string
          enum:
            - "1.0"
        spaces:
          type: array
          items:
            $ref: '#/components/schemas/Space'
        gateways:
          type: array
          items:
            $ref: '#/components/schemas/Gates'
        providers:
          type: array
          items:
            $ref: '#/components/schemas/Provider'
      required:
        - filetype
        - version

    ConnectionRequest:
      type: object
      properties:
        time:
          type: integer
        gateid:
          type: string
        spaceid:
          type: string
        wg_public:
          type: string
        days:
          type: integer
        sessionid:
          type: string
      required:
        - gateid
        - spaceid
        - days

    ConnectionInfo:
      type: object
      properties:
        time:
          type: integer
        username:
          type: string
        password:
          type: string
        bearer:
          type: string
        wallet:
          type: string
        price:
          type: number
        paymentid:
          type: string
        wg_ip:
          type: string
        wg_network_bits:
          type: integer
        wg_gw:
          type: string
        wg_dns:
          type: string
        sessionid:
          type: string
        expires:
          type: integer
      required:
        - time
        - sessionid
        - paymentid
        - wallet
        - username
        - password
        - price
        - expires
        - bearer